fix: unify sandboxed iframe bridge, fix localStorage SecurityError in preview iframes
Root cause: app preview thumbnail iframes (app-card.tsx) rendered untrusted
HTML with sandbox='allow-scripts' but without the storage polyfill that other
sandboxed iframes inject. Framework code (React, etc.) accessing localStorage
during initialization threw SecurityError, crashing the preview.

Changes:
- New utils/sandbox-bridge.ts: single source of truth for all sandboxed iframe
  utilities (storage polyfill, bridge script, safe HTML injection, preview prep)
- New hooks/use-sandbox-fetch-proxy.ts: shared hook for postMessage-based fetch
  proxying, replacing ~80 lines of duplicated handler code
- Unified message protocol uses 'frameId' instead of split 'appId'/'surfaceId'
- app-card.tsx: now injects storage polyfill via preparePreviewHtml()
- app-viewer-container.tsx: uses shared injectBridge() + useSandboxFetchProxy()
- dynamic-page-surface.tsx: deleted ~140 lines of local bridge code, uses shared
  modules. Gains lastIndexOf injection fix and jsonForScript safety escaping.
- Deleted old app-bridge.ts, consolidated tests into sandbox-bridge.test.ts
