+++ apps/web/src/utils/sandbox-bridge.ts
@@ -1,0 +1,236 @@
+/**
+ * Sandboxed iframe bridge utilities.
+ *
+ * Provides everything needed to safely render untrusted HTML inside
+ * `<iframe sandbox="allow-scripts">` (no `allow-same-origin`):
+ *
+ * 1. **Storage polyfill** — in-memory shim for `localStorage` / `sessionStorage`,
+ *    which throw `SecurityError` in sandboxed contexts without `allow-same-origin`.
+ * 2. **Action bridge** — `window.vellum.sendAction()` forwards surface actions to
+ *    the parent via `postMessage`.
+ * 3. **Fetch proxy** — `window.vellum.fetch()` proxies authenticated requests
+ *    through the parent, keeping auth tokens out of the sandbox.
+ * 4. **Safe injection** — `injectScript()` inserts `<script>` tags using
+ *    `lastIndexOf` to avoid hijacking when app JS contains literal close-tag
+ *    sequences.
+ *
+ * All sandboxed iframes that render untrusted HTML must use these utilities.
+ * The storage polyfill is required even for non-interactive preview iframes.
+ *
+ * @see https://developer.mozilla.org/en-US/docs/Web/HTML/Element/iframe#sandbox
+ * @see https://html.spec.whatwg.org/multipage/iframe-embed-object.html#attr-iframe-sandbox
+ */
+
+// ---------------------------------------------------------------------------
+// Constants
+// ---------------------------------------------------------------------------
+
+/** Path pattern allowed through the fetch proxy (matches desktop ATL-83 restriction). */
+export const FETCH_PROXY_PATH_RE = /^\/v1\/x\//;
+
+// ---------------------------------------------------------------------------
+// Script serialization
+// ---------------------------------------------------------------------------
+
+/**
+ * Safely serialize a value for embedding inside an inline `<script>` block.
+ *
+ * `JSON.stringify` alone doesn't escape `</script>` or `<!--`, which can
+ * break out of the script context in `srcdoc`. We replace the two dangerous
+ * sequences after stringifying.
+ *
+ * @see https://html.spec.whatwg.org/multipage/scripting.html#restrictions-for-contents-of-script-elements
+ */
+export function jsonForScript(value: unknown): string {
+  return JSON.stringify(value).replace(/<\//g, "<\\/").replace(/<!--/g, "<\\!--");
+}
+
+// ---------------------------------------------------------------------------
+// Storage polyfill
+// ---------------------------------------------------------------------------
+
+/**
+ * Build a `<script>` tag that replaces `localStorage` and `sessionStorage`
+ * with in-memory shims.
+ *
+ * Required for every sandboxed iframe without `allow-same-origin`. Without
+ * this, any code that accesses `window.localStorage` — including framework
+ * initialization in React, Vue, i18next, etc. — throws a `SecurityError`
+ * and prevents the app from mounting.
+ */
+export function buildStoragePolyfill(): string {
+  return `<script>
+(function() {
+  var store = {};
+  var storageShim = {
+    getItem: function(k) { return store.hasOwnProperty(k) ? store[k] : null; },
+    setItem: function(k, v) { store[k] = String(v); },
+    removeItem: function(k) { delete store[k]; },
+    clear: function() { store = {}; },
+    get length() { return Object.keys(store).length; },
+    key: function(i) { return Object.keys(store)[i] || null; }
+  };
+  try {
+    Object.defineProperty(window, 'localStorage', { value: storageShim, writable: true, configurable: true });
+  } catch(e) { window.localStorage = storageShim; }
+  try {
+    Object.defineProperty(window, 'sessionStorage', { value: storageShim, writable: true, configurable: true });
+  } catch(e) { window.sessionStorage = storageShim; }
+})();
+</script>`;
+}
+
+// ---------------------------------------------------------------------------
+// Full bridge script
+// ---------------------------------------------------------------------------
+
+export interface BridgeOptions {
+  /** Include the authenticated fetch proxy bridge. Default: false. */
+  fetch?: boolean;
+  /** Deep-link route exposed as `window.vellum.route`. */
+  route?: string;
+}
+
+/**
+ * Build the full JS bridge script injected into sandboxed app iframes.
+ *
+ * Includes the storage polyfill, action sender, and optionally the fetch
+ * proxy. The `frameId` is embedded in all `postMessage` payloads so the
+ * parent can route messages when multiple sandboxed iframes coexist.
+ */
+export function buildBridgeScript(frameId: string, options?: BridgeOptions): string {
+  const enableFetch = options?.fetch ?? false;
+  const route = options?.route ?? null;
+
+  const fetchBridge = enableFetch
+    ? `
+  window.vellum._pendingFetches = {};
+  window.vellum._fetchNextId = 1;
+  window.vellum._resolveFetch = function(callId, status, statusText, body, headers) {
+    var p = window.vellum._pendingFetches[callId];
+    if (!p) return;
+    delete window.vellum._pendingFetches[callId];
+    p.resolve({
+      ok: status >= 200 && status < 300,
+      status: status,
+      statusText: statusText,
+      headers: headers || {},
+      _body: body,
+      json: function() { return Promise.resolve(JSON.parse(body)); },
+      text: function() { return Promise.resolve(body); }
+    });
+  };
+  window.vellum._rejectFetch = function(callId, errorMessage) {
+    var p = window.vellum._pendingFetches[callId];
+    if (!p) return;
+    delete window.vellum._pendingFetches[callId];
+    p.reject(new Error(errorMessage));
+  };
+  window.addEventListener('message', function(event) {
+    var d = event.data;
+    if (!d) return;
+    if (d.type === 'vellum_fetch_response' && d.callId) {
+      if (d.error) {
+        window.vellum._rejectFetch(d.callId, d.error);
+      } else {
+        window.vellum._resolveFetch(d.callId, d.status, d.statusText, d.body, d.headers);
+      }
+    }
+  });
+  window.vellum.fetch = function(path, options) {
+    options = options || {};
+    return new Promise(function(resolve, reject) {
+      var callId = 'f' + (window.vellum._fetchNextId++);
+      window.vellum._pendingFetches[callId] = { resolve: resolve, reject: reject };
+      window.parent.postMessage({
+        type: 'vellum_fetch_request',
+        frameId: ${jsonForScript(frameId)},
+        callId: callId,
+        path: path,
+        method: (options.method || 'GET').toUpperCase(),
+        headers: options.headers || {},
+        body: options.body || null
+      }, '*');
+    });
+  };`
+    : "";
+
+  return `<script>
+(function() {
+  var store = {};
+  var storageShim = {
+    getItem: function(k) { return store.hasOwnProperty(k) ? store[k] : null; },
+    setItem: function(k, v) { store[k] = String(v); },
+    removeItem: function(k) { delete store[k]; },
+    clear: function() { store = {}; },
+    get length() { return Object.keys(store).length; },
+    key: function(i) { return Object.keys(store)[i] || null; }
+  };
+  try {
+    Object.defineProperty(window, 'localStorage', { value: storageShim, writable: true, configurable: true });
+  } catch(e) { window.localStorage = storageShim; }
+  try {
+    Object.defineProperty(window, 'sessionStorage', { value: storageShim, writable: true, configurable: true });
+  } catch(e) { window.sessionStorage = storageShim; }
+
+  window.vellum = {
+    route: ${jsonForScript(route)},
+    sendAction: function(actionId, data) {
+      window.parent.postMessage({
+        type: 'vellum_surface_action',
+        frameId: ${jsonForScript(frameId)},
+        actionId: actionId,
+        data: data || {}
+      }, '*');
+    }
+  };${fetchBridge}
+})();
+</script>`;
+}
+
+// ---------------------------------------------------------------------------
+// HTML injection
+// ---------------------------------------------------------------------------
+
+/**
+ * Safely inject a script string into HTML.
+ *
+ * Insertion priority: before the last `</body>`, then after the last
+ * `</head>`, then prepended. Uses `lastIndexOf` so that literal close-tag
+ * sequences inside `<script>` blocks (comments, strings) can't hijack the
+ * injection site.
+ */
+export function injectScript(html: string, script: string): string {
+  const BODY_CLOSE = "</body>";
+  const HEAD_CLOSE = "</head>";
+
+  const bodyIdx = html.lastIndexOf(BODY_CLOSE);
+  if (bodyIdx !== -1) {
+    return html.slice(0, bodyIdx) + script + html.slice(bodyIdx);
+  }
+  const headIdx = html.lastIndexOf(HEAD_CLOSE);
+  if (headIdx !== -1) {
+    const after = headIdx + HEAD_CLOSE.length;
+    return html.slice(0, after) + script + html.slice(after);
+  }
+  return script + html;
+}
+
+/**
+ * Inject the full bridge into app HTML.
+ */
+export function injectBridge(html: string, frameId: string, options?: BridgeOptions): string {
+  return injectScript(html, buildBridgeScript(frameId, options));
+}
+
+/**
+ * Prepare HTML for a non-interactive preview iframe.
+ *
+ * Injects the storage polyfill (prevents `SecurityError` when app code
+ * accesses `localStorage` during initialization) and hides scrollbars.
+ */
+export function preparePreviewHtml(html: string): string {
+  const HIDE_SCROLLBARS =
+    "<style>html,body{overflow:hidden!important;scrollbar-width:none!important;}::-webkit-scrollbar{display:none!important;}</style>";
+  return injectScript(html, buildStoragePolyfill() + HIDE_SCROLLBARS);
+}
